# Optimize the text measuring of legend items when rendering large amount of traces