I open this as a separate PR from https://github.com/plotly/plotly.js/pull/1585, because I am not sure if you guys will be willing to take this one.

The logic behind my change here is that if you want to plot reasonable amount of traces, we will give one very precise measuring and layout on legend items. But if you go crazy, we will sacrifice the precision of legend item layout (which probably does not matter anymore in this case) for way much better performance.

On my machine, this change reduced the rendering time for
- 100 traces (200 points per trace) from 1.6 seconds to 1.1 seconds
- 300 traces (200 points per trace) from 5.7 seconds to 2.8 seconds
- 500 traces (200 points per trace) from 12.4 seconds to 5.8 seconds