@etpinard 2017-04-12
My gut feeling would be to :-1: this PR. I don't think the perf improvement here is worth the maintenance burden / inconsistency. 

Perhaps instead, we should make `layout.showlegend` default to `false` when there are more then say 100 traces on the graph in v2?